# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blog API",
    "description": "\n## 博客系统 API\n\n一个功能完整的现代化博客系统后端 API，支持 Markdown/MDX 内容管理和 GitOps 自动化同步。\n\n### 核心特性\n\n- 🔐 **用户认证**：基于 JWT 的安全认证系统\n- 📝 **内容管理**：支持 Markdown/MDX 格式的文章创建和编辑\n- 🎨 **三种渲染模式**：后端渲染、前端 SSR、前端 CSR\n- 📁 **媒体管理**：图片、视频等媒体文件的上传和管理\n- 🔄 **GitOps 同步**：Git 仓库与数据库的双向自动化同步\n- 🔗 **Webhook 集成**：支持 GitHub Webhook 自动触发同步\n\n### 技术栈\n\n- **框架**：FastAPI + SQLModel + PostgreSQL\n- **认证**：JWT + OAuth2\n- **内容处理**：markdown-it-py + MDX\n- **异步支持**：asyncio + aiofiles\n\n### 快速开始\n\n1. 注册账号：`POST /api/v1/users/register`\n2. 登录获取 token：`POST /api/v1/users/login`\n3. 使用 token 访问受保护的接口：`Authorization: Bearer <token>`\n\n### 文档\n\n- [Swagger UI](/docs) - 交互式 API 文档\n- [Scalar UI](/scalar) - 现代化 API 文档\n- [OpenAPI JSON](/openapi.json) - OpenAPI 规范文件\n    ",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/posts/{post_type}/{post_id}/like": {
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "点赞文章",
        "description": "点赞文章（增加点赞数）\n\n## 权限\n- 无需认证，任何人都可以点赞\n- 不记录点赞用户，仅统计点赞总数\n\n## 路径参数\n- `post_type`: 板块类型（article/idea）\n- `post_id`: 文章 UUID\n\n## 返回值\n```json\n{\n    \"like_count\": 42\n}\n```\n\n## 示例\n```bash\n# 点赞文章\nPOST /posts/article/123e4567-e89b-12d3-a456-426614174000/like\n\n# 响应\n{\n    \"like_count\": 43\n}\n```\n\n## 错误码\n- `404 NOT_FOUND`: 文章不存在\n- `500 INTERNAL_SERVER_ERROR`: 数据库更新失败\n\n## 注意事项\n- 不验证用户身份，允许重复点赞\n- 点赞数不会减少到负数\n- 建议前端实现防抖，避免重复请求",
        "operationId": "likePost",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "文章 ID",
              "title": "Post Id"
            },
            "description": "文章 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostLikeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "posts"
        ],
        "summary": "取消点赞",
        "description": "取消点赞（减少点赞数）\n\n## 权限\n- 无需认证，任何人都可以取消点赞\n- 不记录点赞用户，仅统计点赞总数\n\n## 路径参数\n- `post_type`: 板块类型（article/idea）\n- `post_id`: 文章 UUID\n\n## 返回值\n```json\n{\n    \"like_count\": 41\n}\n```\n\n## 示例\n```bash\n# 取消点赞\nDELETE /posts/article/123e4567-e89b-12d3-a456-426614174000/like\n\n# 响应\n{\n    \"like_count\": 42\n}\n```\n\n## 错误码\n- `404 NOT_FOUND`: 文章不存在\n- `500 INTERNAL_SERVER_ERROR`: 数据库更新失败\n\n## 注意事项\n- 点赞数不会减少到负数（最小为 0）\n- 建议前端实现防抖，避免重复请求",
        "operationId": "unlikePost",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "文章 ID",
              "title": "Post Id"
            },
            "description": "文章 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostLikeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/{post_type}/{post_id}/bookmark": {
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "收藏文章",
        "description": "收藏文章（增加收藏数）\n\n## 权限\n- 无需认证，任何人都可以收藏\n- 不记录收藏用户，仅统计收藏总数\n\n## 路径参数\n- `post_type`: 板块类型（article/idea）\n- `post_id`: 文章 UUID\n\n## 返回值\n```json\n{\n    \"bookmark_count\": 15\n}\n```\n\n## 示例\n```bash\n# 收藏文章\nPOST /posts/article/123e4567-e89b-12d3-a456-426614174000/bookmark\n\n# 响应\n{\n    \"bookmark_count\": 16\n}\n```\n\n## 错误码\n- `404 NOT_FOUND`: 文章不存在\n- `500 INTERNAL_SERVER_ERROR`: 数据库更新失败\n\n## 注意事项\n- 不验证用户身份，允许重复收藏\n- 收藏数不会减少到负数\n- 建议前端实现防抖，避免重复请求\n- 如需实现用户收藏列表，需要额外的用户收藏关系表",
        "operationId": "bookmarkPost",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "文章 ID",
              "title": "Post Id"
            },
            "description": "文章 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostBookmarkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "posts"
        ],
        "summary": "取消收藏",
        "description": "取消收藏（减少收藏数）\n\n## 权限\n- 无需认证，任何人都可以取消收藏\n- 不记录收藏用户，仅统计收藏总数\n\n## 路径参数\n- `post_type`: 板块类型（article/idea）\n- `post_id`: 文章 UUID\n\n## 返回值\n```json\n{\n    \"bookmark_count\": 14\n}\n```\n\n## 示例\n```bash\n# 取消收藏\nDELETE /posts/article/123e4567-e89b-12d3-a456-426614174000/bookmark\n\n# 响应\n{\n    \"bookmark_count\": 15\n}\n```\n\n## 错误码\n- `404 NOT_FOUND`: 文章不存在\n- `500 INTERNAL_SERVER_ERROR`: 数据库更新失败\n\n## 注意事项\n- 收藏数不会减少到负数（最小为 0）\n- 建议前端实现防抖，避免重复请求",
        "operationId": "unbookmarkPost",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "文章 ID",
              "title": "Post Id"
            },
            "description": "文章 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostBookmarkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "timestamp",
          "request_id"
        ],
        "title": "ErrorDetail",
        "description": "错误详情模型"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "统一错误响应模型"
      },
...
      "PostBookmarkResponse": {
        "properties": {
          "bookmark_count": {
            "type": "integer",
            "title": "Bookmark Count"
          }
        },
        "type": "object",
        "required": [
          "bookmark_count"
        ],
        "title": "PostBookmarkResponse",
        "description": "收藏响应"
      },
...
      "PostLikeResponse": {
        "properties": {
          "like_count": {
            "type": "integer",
            "title": "Like Count"
          }
        },
        "type": "object",
        "required": [
          "like_count"
        ],
        "title": "PostLikeResponse",
        "description": "点赞响应"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/users/login"
          }
        }
      }
    }
  },
...
}